Refactor `sparkLM` summary method
Created a new `summaryObj` function that will return a named list containing all the pieces of the summary output as well as a `print` function that closes over the the existing list elements. This is so we can pass summary objects over the wire and print them on the client.

`summary.sparkLM` now simply takes the model object, calls `summaryObj` on it, and then calls the print closure.

The result is that you can still call `summary` like normal and have the results print to the console, but you can also pass the summary info to the client and print it there instead.